Trait deathframe::components::component_prelude::ActionQueue

pub trait ActionQueue {
    type Action;
    fn mut_actions(&mut self) -> &mut Vec<Self::Action>;

    fn add_action(&mut self, action: Self::Action) { ... }

    fn drain_actions(&mut self) -> Drain<Self::Action> { ... }
}

The ActionQueue can receive and accumulate actions, which can then be consumed at some point.

Associated Types

type Action

Loading content...
Required methods

fn mut_actions(&mut self) -> &mut Vec<Self::Action>

Returns a mutable reference to the Vec of actions Action.

Loading content...
Provided methods

fn add_action(&mut self, action: Self::Action)

Add an Action to the action queue.

fn drain_actions(&mut self) -> Drain<Self::Action>

Returns a draining iterator over all queued actions. Consume the actions in the queue.

Loading content...