Trait deathframe::resources::prelude::AudioManager

pub trait AudioManager<K> where
    K: PartialEq<K> + Eq + Hash, {
    fn get_source_handle(&self, key: &K) -> Option<&Handle<Source>>;

    fn insert_source_handle(&mut self, key: K, source_handle: Handle<Source>);

    fn load_audio<P>(
        &mut self, 
        key: K, 
        path: P, 
        loader: &Loader, 
        asset_storage: &AssetStorage<Source>
    ) -> Result<(), String>
    where
        P: AsRef<Path>,
    { ... }
}

Trait with common functions between the sound/music managers.

Required methods

fn get_source_handle(&self, key: &K) -> Option<&Handle<Source>>

Returns a reference to the SourceHandle for the given key.

fn insert_source_handle(&mut self, key: K, source_handle: Handle<Source>)

Should insert the given SourceHandle with the given key.

Loading content...
Provided methods

fn load_audio<P>(
&mut self,
key: K,
path: P,
loader: &Loader,
asset_storage: &AssetStorage<Source>
) -> Result<(), String> where
P: AsRef<Path>,

Load sound file from for key K from the given path. The file format is derived from the filename's extension. Returns an Error, if no matching audio format was found for the file extension. Valid extensions: ".wav", ".mp3", ".ogg", ".flac"

Loading content...