Struct deathframe::systems::system_prelude::EntityComponentInserter

pub struct EntityComponentInserter { /* fields omitted */ }

This struct accumulates InsertionActions for entities, and then inserts or removes a certain component from all entities at once, by calling the run method. Component to insert/remove has to have Default implemented. It can prioritize insertion or removal over the other action. Used with EntityLoaderSystem.

Methods

impl EntityComponentInserter

pub fn with_priority(
self,
prioritize_action: InsertionAction
) -> EntityComponentInserter

Which InsertionAction to prioritize. So if the Insert action is prioritized, then no subsequent Remove action can ever overwrite the Insert action. And vice versa. See InsertionAction's Default implementation for the default.

pub fn with_cache(self, use_cache: bool) -> EntityComponentInserter

If caching logic should be used. The caching is pretty bad, and seems to slow things down, rather than speed stuff up. Not recommended.

pub fn insert(&mut self, entity: Entity)

Stage the given entity for insertion. So on run, the given component will be added to the entity.

pub fn remove(&mut self, entity: Entity)

Stage the given entity for removal. So on run, the given component will be removed from the entity.

pub fn run<C>(
&mut self,
storage: &mut Storage<C, FetchMut<MaskedStorage<C>>>
) -> Result<(), Error> where
C: Component + Default,

Run all InsertionActions at once. When inserting, inserts the Default component of the given component storage.